Solution - Section 4 Exercise - Add a form to create/edit a movie.

+++ Models/Movie.cs
@@ -26,7 +26,7 @@ namespace Vidly.Models
 
         public DateTime ReleaseDate { get; set; }
 
-        public byte NumberInStock { get; set; }
+        public int NumberInStock { get; set; }
 
     }
 }
